Move internal test command to a separate package
