+++ internal/commands/internal/internal.go
@@ -1,6 +1,7 @@
 package internal
 
 import (
+	"github.com/cirruslabs/cirrus-cli/internal/commands/internal/test"
 	"github.com/spf13/cobra"
 )
 
@@ -12,7 +13,7 @@ func NewRootCmd() *cobra.Command {
 	}
 
 	cmd.AddCommand(
-		newTestCmd(),
+		test.NewTestCmd(),
 	)
 
 	return cmd
