+++ internal/commands/internal/test/testdata/simple/dir/subdir/.cirrus.star
