+++ internal/commands/internal/test/testdata/simple/lib.star
